requirements.txt
configparser>=3.5.0
Jinja2>=2.10
python-dateutil>=2.7.5
ratelimit>=2.2.0
requests>=2.21.0
coloredlogs>=10.0
argcomplete>=1.9.4
markdown2>=2.3.7
GitPython>=2.1.11
bjoern>=2.2.3
flask>=1.0.2
gevent>=1.4.0
flask-restplus>=0.12.1
SQLAlchemy>=1.3.1
Flask-SQLAlchemy>=2.4.0
Alembic==0.8.10
werkzeug>=0.14.1,<0.15

marshmallow>=2.13.5
flask-marshmallow==0.7.0
marshmallow-sqlalchemy==0.12.0
webargs>=1.4.0
apispec>=0.20.0,<0.39

docker
verboselogs>=1.7
beautifulsoup4
psutil>=5.5.1
PyYAML
pymdown-extensions
pytest>=4.4.1

flask-caching>=1.7.1